Refactoring PR
