+++ metadata-io/build.gradle
@@ -13,6 +13,7 @@ dependencies {
   compile project(':metadata-events:mxe-registration')
   compile project(':metadata-events:mxe-utils-avro-1.7')
   compile project(path: ':metadata-models')
+  compile project(':metadata-service:restli-client')
 
   compile spec.product.pegasus.data
   compile spec.product.pegasus.generator
