+++ metadata-service/auth-impl/build.gradle
@@ -7,8 +7,8 @@ compileJava {
 
 dependencies {
   compile project(path: ':metadata-models')
-  compile project(path: ':metadata-service:restli-client')
   compile project(path: ':metadata-service:auth-api')
+  compile project(path: ':metadata-io')
 
   implementation 'io.jsonwebtoken:jjwt-api:0.11.2'
   runtimeOnly 'io.jsonwebtoken:jjwt-impl:0.11.2',
