+++ metadata-service/factories/src/main/java/com/linkedin/gms/factory/auth/GroupServiceFactory.java
@@ -3,7 +3,7 @@
 package com.linkedin.gms.factory.auth;
 
 import com.datahub.authentication.group.GroupService;
-import com.linkedin.entity.client.JavaEntityClient;
+import com.linkedin.metadata.client.JavaEntityClient;
 import com.linkedin.gms.factory.spring.YamlPropertySourceFactory;
 import com.linkedin.metadata.entity.EntityService;
 import com.linkedin.metadata.graph.GraphClient;
