+++ metadata-service/restli-client/build.gradle
@@ -3,11 +3,9 @@ apply plugin: 'pegasus'
 dependencies {
   compile project(':metadata-service:restli-api')
   compile project(':metadata-service:auth-api')
-  compile project(':metadata-service:restli-servlet-impl')
   compile project(path: ':metadata-service:restli-api', configuration: 'restClient')
   compile project(':metadata-events:mxe-schemas')
-
-  dataModel project(':li-utils')
+  compile project(':metadata-utils')
 
   compile externalDependency.lombok
   annotationProcessor externalDependency.lombok
